issue where HDF5 particle ID array was being returned as a float32 because of the 1.0 conversion multiplication - should not return unsigned int array.

+++ readgadget/modules/hdf5.py
@@ -36,7 +36,8 @@ def hdf5_general(f,h,ptype):
         if h.units and h.reading == 'u':
             ne = f['PartType0/%s' % (HDF5_NAMES['ne'])]
             import common as common
-            h.convert = common.getTfactor(np.asarray(ne),h)
+            h.convert = common.getTfactor(np.asarray(ne,dtype=np.float32),h)
+            h.convert = h.convert.astype(np.float32)
     else:
         if h.debug: print 'coult not find PartType%d' % ptype
         arr = np.zeros(0,dtype=np.float32)
